Removed use of deprecated wxTE_AUTO_SCROLL.
Use of this macro prevents compilation with wxWidgets that has been
compiled with 2.8 compatibility turned off.  This macro is set to 0
and the comment in the wxWidgets source reads:

this style is (or at least should be) on by default now, don't use it

This commit removed the use of this obsolete macro.