Fix code generation when multiple scenes used the same events.
Code generation was modifying the events in place. If two or more scenes
used the same events, the events could be modified for a scene during code
generation, leading to a bad code generation for the others.

(In particular, if a scene was not containing an object, any action related to it
would be set an action with an empty type. Any other scene using the action will then
have the action not working).