English Improvement+ fixed dead link + new message
This link appears in GDevelop, but it's dead:
http://compilgames.net/wiki/doku.php/game_develop/extensions

In an old discussion on your forum, Victor said the old link should be changed to this one:
http://wiki.compilgames.net/doku.php/gdevelop/extensions

This is the discussion:
http://www.forum.compilgames.net/viewtopic.php?f=19&t=7580

That page is empty because people haven't added extensions yet, so I changed the message that appears.

This is the old message:
"A list of extensions provided by other developers is available on the wiki"

This is the new message:
"If you create new extensions for GDevelop, please share them here"

I am afraid of the old message causing people to become hopeful, and then they'll be disappointed when they see the page is empty, so my new message seems better. Tell me if you're unhappy with my change.

I also translated the Wiki page, so it exists in the English wiki now. Victor's extensions appear on the page, but the links are dead. Should I delete them from the list? I'm afraid of people being hopeful and then disappointed:
http://wiki.compilgames.net/doku.php/gdevelop/extensions