GDJS: store global volume value in HowlerSoundManager
Every programmer knows why 0.01 - 0.01 may not equal 0.
However, not eveyone knows why 1 - 1 may not equal 0 :)

As GDevelop uses values 0-100 for its global volume, it's easy
to encounter floating point inaccuracy problems while doing trivial
things like a loop that decrements 1 from the global volume.

I think a chance for a non-programmer to figure out why 1 - 1 is
not 0 is actually 0, so let's work on integers to remove the confussion.