fix bug  for cancelled piskl edits + lock gdevelop while piskel is open
this.animationEdited  also needs to be set to false if piskel changes were cancelled, otherwise the next piskel edit would write on it - even if its for another animation clip.
Also locked mouse events on gdevelop while piskel's window is open- to prevent users from changing the order of animations in gdevelop while editing in piskel.