Add ability to rename object from object editor
This addresses these usability problems:
- Newly created objects can now be immediately named- creating a better rapid prototyping flow
- Currently edited objects show their name in the object editor

This checks if the object can use the new name when the input box loses focus. It applies the new name when the user actually clicks the apply button

TODO: possibly refactor the renaming of object  in the object list to use the checking function that I added to the index - should simplify code.
Will do that after getting some feedback