Add watcher to automatically import changes to GDJS/extensions in newIDE
This watches for changes in GDJS or extensions and automatically run the "import-GDJS-Runtime.js" script.

This allows for faster iterations when working on GDJS or JS extensions: do some changes, verify that the console says that changes were imported, launch the preview

Adds an option in preferences to deactivate this in case of issues.

Also move BrowserS3GDJSFinder and LocalGDJSFinder in newIDE/app/src/GameEngineFinder.