Move LocalFolderPicker on top
This LocalFolderPicker  should be the first thing to setup. Then we can choose which example choose.
Otherwise if you click on an example first, the project is saved at the default location, this behavior grow up the size of the default folder. And after some years of use you can see the folder "My project 416948984".
If the user see in first where the folder is save it's really proper than the current behavior.