Behavior change, grabbing look behind instead of looking ahead
There is a one step difference in the tests. But I think it won't mater because the grab interval is still the same when falling.
The interval when jumping was really small comparing to coords delta. That why it didn't grab but still could. This is fixed with an explicite condition.
One test strengthen.