RuntimeScenePixiRenderer: Wait with showing the cursor until the frame is rendered
This minimizes cursor flicker in games that switch between system cursor and
in-game cursor. If the browser updates cursor status asynchronously and the
rendering takes longer than screen refresh, we could easily end up with two
cursors displayed at once for a short moment.

While the same could be done when hiding the cursor as well, it may be not
a bad idea to do the hiding early anyway, as system cursor can be updated
faster than the browser is able to dispatch events, so hiding it early
minimizes the cases where the cursor jumps back to a previous position
after switching to in-game cursor.